Remove rest of configs from `sc-network-common` to `sc-network`
