+++ client/chain-spec/Cargo.toml
@@ -19,7 +19,7 @@ serde_json = "1.0.85"
 sc-client-api = { version = "4.0.0-dev", path = "../api" }
 sc-chain-spec-derive = { version = "4.0.0-dev", path = "./derive" }
 sc-executor = { version = "0.10.0-dev", path = "../executor" }
-sc-network-common = { version = "0.10.0-dev", path = "../network/common" }
+sc-network = { version = "0.10.0-dev", path = "../network" }
 sc-telemetry = { version = "4.0.0-dev", path = "../telemetry" }
 sp-blockchain = { version = "4.0.0-dev", path = "../../primitives/blockchain" }
 sp-core = { version = "7.0.0", path = "../../primitives/core" }
