+++ client/network/src/discovery.rs
@@ -46,7 +46,7 @@
 //! active mechanism that asks nodes for the addresses they are listening on. Whenever we learn
 //! of a node's address, you must call `add_self_reported_address`.
 
-use crate::utils::LruHashSet;
+use crate::{config::ProtocolId, utils::LruHashSet};
 
 use array_bytes::bytes2hex;
 use futures::prelude::*;
@@ -75,7 +75,6 @@ use libp2p::{
 	},
 };
 use log::{debug, info, trace, warn};
-use sc_network_common::config::ProtocolId;
 use sp_core::hexdisplay::HexDisplay;
 use std::{
 	cmp,
@@ -906,6 +905,7 @@ mod tests {
 	use super::{
 		kademlia_protocol_name, legacy_kademlia_protocol_name, DiscoveryConfig, DiscoveryOut,
 	};
+	use crate::config::ProtocolId;
 	use futures::prelude::*;
 	use libp2p::{
 		core::{
@@ -917,7 +917,6 @@ mod tests {
 		swarm::{Executor, Swarm, SwarmEvent},
 		yamux, Multiaddr,
 	};
-	use sc_network_common::config::ProtocolId;
 	use sp_core::hash::H256;
 	use std::{collections::HashSet, pin::Pin, task::Poll};
