+++ client/network/sync/src/service/mock.rs
@@ -21,11 +21,12 @@ use libp2p::{Multiaddr, PeerId};
 
 use sc_consensus::{BlockImportError, BlockImportStatus};
 use sc_network::{
+	config::MultiaddrWithPeerId,
 	request_responses::{IfDisconnected, RequestFailure},
 	NetworkNotification, NetworkPeers, NetworkRequest, NetworkSyncForkRequest,
 	NotificationSenderError, NotificationSenderT,
 };
-use sc_network_common::{config::MultiaddrWithPeerId, protocol::ProtocolName};
+use sc_network_common::protocol::ProtocolName;
 use sc_peerset::ReputationChange;
 use sp_runtime::traits::{Block as BlockT, NumberFor};
