+++ client/network/sync/src/state_request_handler.rs
@@ -30,8 +30,10 @@ use lru::LruCache;
 use prost::Message;
 
 use sc_client_api::{BlockBackend, ProofProvider};
-use sc_network::request_responses::{IncomingRequest, OutgoingResponse, ProtocolConfig};
-use sc_network_common::config::ProtocolId;
+use sc_network::{
+	config::ProtocolId,
+	request_responses::{IncomingRequest, OutgoingResponse, ProtocolConfig},
+};
 use sp_runtime::traits::Block as BlockT;
 
 use std::{
