+++ client/rpc/src/system/tests.rs
@@ -99,7 +99,7 @@ fn api<T: Into<Option<Status>>>(sync: T) -> RpcModule<System<Block>> {
 					);
 				},
 				Request::NetworkAddReservedPeer(peer, sender) => {
-					let _ = match sc_network_common::config::parse_str_addr(&peer) {
+					let _ = match sc_network::config::parse_str_addr(&peer) {
 						Ok(_) => sender.send(Ok(())),
 						Err(s) =>
 							sender.send(Err(error::Error::MalformattedPeerArg(s.to_string()))),
